Migrate web editor from Quill to react-native-enriched
Unify the comment composer on react-native-enriched across all platforms.
Previously native used react-native-enriched (0.5.2) while web used Quill
(react-quill-new); now that react-native-enriched 0.7.0 ships experimental
web support via package `exports` conditions, both platforms share a single
implementation.

- Bump react-native-enriched 0.5.2 -> 0.7.0 (web support); make it a
  required (non-optional) peer at >=0.7.0.
- Drop quill / react-quill-new from dev and peer dependencies.
- Collapse comment-text-area.native.tsx + comment-text-area.web.tsx into a
  single comment-text-area.tsx. The same EnrichedTextInput API resolves to
  the native build on Metro and the web build on web bundlers, so web now
  gets the @mention popup that the Quill path was missing, plus inline-code
  formatting (toggleInlineCode) on both platforms.
- example-web: pre-bundle react-native-enriched instead of quill, and add an
  exact-match react-native -> react-native-web alias so the enriched web
  build's bare `react-native` import (processColor) resolves during the
  production build.
- Update the stale README (it still referenced the long-removed 10tap editor
  and react-native-webview) and the tests-ui enriched mock.
