+++ package.json
@@ -63,12 +63,10 @@
     "metro-react-native-babel-preset": "^0.66.2",
     "pod-install": "^0.1.0",
     "prettier": "^2.0.5",
-    "quill": "^2.0.3",
     "react": "18.3.1",
     "react-native": "^0.69.4",
     "react-native-builder-bob": "^0.18.2",
-    "react-native-enriched": "^0.5.2",
-    "react-quill-new": "^3.8.3",
+    "react-native-enriched": "^0.7.0",
     "react-test-renderer": "^18.3.1",
     "ts-clean": "^1.0.3",
     "ts-jest": "^29.0.1",
@@ -85,15 +83,6 @@
   "peerDependencies": {
     "react": "*",
     "react-native": "*",
-    "react-native-enriched": ">=0.4.0",
-    "react-quill-new": ">=3.8.0"
-  },
-  "peerDependenciesMeta": {
-    "react-native-enriched": {
-      "optional": true
-    },
-    "react-quill-new": {
-      "optional": true
-    }
+    "react-native-enriched": ">=0.7.0"
   }
 }
